+++ lib/src/widgets/query.dart
@@ -49,7 +49,7 @@ class _QueryState<T> extends State<Query> {
     return StreamBuilder<QueryResponse<T>>(
       initialData: QueryResponse<T>(
         queryRequest: widget.queryRequest,
-        source: ResponseSource.None,
+        dataSource: DataSource.None,
       ),
       stream: stream,
       builder: (context, snapshot) => builder(